requirements.txt
streamlit==1.32.0
openai==1.34.0
python-dotenv==1.0.1
